added saml mapper. refactored into a common interface
